+++ lib/ts/recipe/session/sessionClass.ts
@@ -43,7 +43,7 @@ export default class Session implements SessionContainerInterface {
     }
 
     revokeSession = async (userContext?: any) => {
-        await this.helpers.sessionRecipeImpl.revokeSession({
+        await this.helpers.getRecipeImpl().revokeSession({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
         });
@@ -58,7 +58,7 @@ export default class Session implements SessionContainerInterface {
     };
 
     getSessionData = async (userContext?: any): Promise<any> => {
-        let sessionInfo = await this.helpers.sessionRecipeImpl.getSessionInformation({
+        let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
         });
@@ -74,7 +74,7 @@ export default class Session implements SessionContainerInterface {
 
     updateSessionData = async (newSessionData: any, userContext?: any) => {
         if (
-            !(await this.helpers.sessionRecipeImpl.updateSessionData({
+            !(await this.helpers.getRecipeImpl().updateSessionData({
                 sessionHandle: this.sessionHandle,
                 newSessionData,
                 userContext: userContext === undefined ? {} : userContext,
@@ -88,7 +88,7 @@ export default class Session implements SessionContainerInterface {
         }
     };
 
-    getUserId = () => {
+    getUserId = (_userContext?: any) => {
         return this.userId;
     };
 
@@ -105,7 +105,7 @@ export default class Session implements SessionContainerInterface {
     };
 
     updateAccessTokenPayload = async (newAccessTokenPayload: any, userContext?: any) => {
-        let response = await this.helpers.sessionRecipeImpl.regenerateAccessToken({
+        let response = await this.helpers.getRecipeImpl().regenerateAccessToken({
             accessToken: this.getAccessToken(),
             newAccessTokenPayload,
             userContext: userContext === undefined ? {} : userContext,
@@ -136,7 +136,7 @@ export default class Session implements SessionContainerInterface {
     };
 
     getTimeCreated = async (userContext?: any): Promise<number> => {
-        let sessionInfo = await this.helpers.sessionRecipeImpl.getSessionInformation({
+        let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
         });
@@ -151,7 +151,7 @@ export default class Session implements SessionContainerInterface {
     };
 
     getExpiry = async (userContext?: any): Promise<number> => {
-        let sessionInfo = await this.helpers.sessionRecipeImpl.getSessionInformation({
+        let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
         });